Broke map.

+++ src/graphics/sprite/Mesh.cpp
@@ -99,7 +99,7 @@ short MeshManager::load( GLBrush* brush )
 
 	Mesh* m = NULL;
 	ITER_LIST( Mesh*, Meshes ){
-		if( it->data->location == brush->mesh )
+		if( (int)it->data->location == brush->mesh )
 			m = it->data;
 	}
